fixing visual studio warnings

+++ CobaltLib/src/Problem.cpp
@@ -166,7 +166,7 @@ CobaltStatus Problem::validate( ) {
     } else { // match B
       // find this index in B
       bool indexFound = false;
-      unsigned int indexB = -1;
+      unsigned int indexB = static_cast<unsigned int>(-1);
       for (unsigned int j = 0; j < tensorB.numDims(); j++) {
         if (indicesB[j] == indexAssignment) {
           indexFound = true;
@@ -195,7 +195,7 @@ CobaltStatus Problem::validate( ) {
     } else { // match A
       // find this index in A
       bool indexFound = false;
-      unsigned int indexA = -1;
+      unsigned int indexA = static_cast<unsigned int>(-1);
       for (unsigned int j = 0; j < tensorA.numDims(); j++) {
         if (indicesA[j] == indexAssignment) {
           indexFound = true;
